tests: fail on unknown error
Ensure tests fail when they exit with an unknown error, such as when a
SIGSEGV occurs. We want the test run to error out when this occurs.

Added some fflush() calls to ensure logs are written and the test name
is printed when there is an unknown error like this.